CDAF 0.8.4 WT-100 check that string is long enough to test for CDAF execute keyword

+++ automation/remote/execute.ps1
@@ -66,15 +66,16 @@ Foreach ($line in get-content $TASK_LIST) {
         # Do not attempt any processing when a line is just a comment
         if ($expression) {
 
-	        # Check for cross platform key words
-            if ( $expression.substring(0,6) -match 'assign' ) {
-	            $expression = $expression.Substring(7)
-            }
-
-	        # Check for cross platform key words
-            if ( $expression.substring(0,6) -match 'remove' ) {
-	            $expression = "rm " + $expression.Substring(7)
-            }
+	        # Check for cross platform key words, only if the string is longer enough
+	        if ($dbUserName.length -gt 6) {
+	            if ( $expression.substring(0,6) -match 'assign' ) {
+		            $expression = $expression.Substring(7)
+	            }
+	
+	            if ( $expression.substring(0,6) -match 'remove' ) {
+		            $expression = "rm " + $expression.Substring(7)
+	            }
+	        }
 
 	        # Do not echo line if it is an echo itself
             if (-not ($expression -match 'Write-Host')) {
